Make helper for replace_in-buffer
